requirements_server.txt
python-dotenv
google-genai
websockets
fastapi
uvicorn
Pillow
numpy
scipy
opencv-python